+++ database/migrations/2025_02_21_154247_create_push_subscriptions_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
 
 class CreatePushSubscriptionsTable extends Migration
 {
@@ -13,15 +14,17 @@ class CreatePushSubscriptionsTable extends Migration
      */
     public function up()
     {
-        Schema::connection(config('webpush.database_connection'))->create(config('webpush.table_name'), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->morphs('subscribable');
-            $table->string('endpoint', 500)->unique();
-            $table->string('public_key')->nullable();
-            $table->string('auth_token')->nullable();
-            $table->string('content_encoding')->nullable();
-            $table->timestamps();
-        });
+        if (DB::getDriverName() === 'pgsql') {
+            Schema::connection(config('webpush.database_connection'))->create(config('webpush.table_name'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->morphs('subscribable');
+                $table->string('endpoint', 500)->unique();
+                $table->string('public_key')->nullable();
+                $table->string('auth_token')->nullable();
+                $table->string('content_encoding')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +34,8 @@ class CreatePushSubscriptionsTable extends Migration
      */
     public function down()
     {
-        Schema::connection(config('webpush.database_connection'))->dropIfExists(config('webpush.table_name'));
+        if (DB::getDriverName() === 'pgsql') {
+            Schema::connection(config('webpush.database_connection'))->dropIfExists(config('webpush.table_name'));
+        }
     }
 }
